// frontend/src/types/api.ts
export interface AnswerRequest {
  query: string;
  top_k: number;
  ef_search: number;
}